Add config.vm.box_url to download Precise 32
Currently if you do not have the precise32 box already downloaded on your computer, vip-quickstart will not run giving the error that the box could not be found. If you already have "precise32" it will not be downloaded again, but those new to vagrant or do not have the precise32 box will not have to find the box and install it themselves.